Consolidate build logic into singe build Move convention utility function
