+++ service/api/build.gradle.kts
@@ -1,4 +1,4 @@
-import buildlogic.repoRoot
+import viaduct.gradle.internal.repoRoot
 
 plugins {
     `java-library`
